refactor(type-guards): move to lib
